pyproject.toml
# SPDX-FileCopyrightText: 2023 Alec Delaney
# SPDX-License-Identifier: MIT

[project]
name = "maybe-not"
description = "Randomly close an issue or pull request when opened"
version = "1.0.0"
requires-python = ">=3.13.0"
readme = "README.md"
authors = [
    {name = "Alec Delaney", email = "tekktrik@gmail.com"}
]
urls = {Homepage = "https://github.com/FastAndCheap/maybe-not"}
license = "MIT"
dependencies = [
    "pygithub>=2.6.1",
]

[tool.ruff.lint]
select = ["D", "PL", "I"]
ignore = ["D213", "D203"]
